% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{asinsqrt.t}
\alias{asinsqrt.t}
\title{Arc sine and square root-transformed test statistic.}
\usage{
asinsqrt.t(surv_KM, se_KM)
}
\arguments{
\item{surv_KM}{a numeric vector containing survival rates for two groups}

\item{se_KM}{a numeric vector containing standard errors for two groups}
}
\value{
A numeric scalar containing the value of the test statistic.
}
\description{
This function is a test statistic for comparing the survival rates of
two survival curves at a fixed point in time, using an arc-sine square-root
transformation of the survival function. It is an implementation of the
test statistic \eqn{X_4^2}{X4^2} in Klein et al.'s paper [1].
}
\examples{
data(exp_surv)

surv_KM = get_surv_KM(exp_surv, t=1)
se_KM = get_se_KM(exp_surv, t=1)
asinsqrt.t(surv_KM, se_KM)

}
\references{
[1]  Klein, J.P., Logan, B., Harhoff, M. and Andersen, P.K. (2007),
\href{https://doi.org/10.1002/sim.2864}{Analyzing survival curves at a fixed point in time}.
Statist. Med., 26: 4505-4519.
}
