% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{clog.test}
\alias{clog.test}
\title{Clog-transformed test for difference of survival rates}
\usage{
clog.test(data = NULL, surv_KM = NULL, se_KM = NULL, t = NULL)
}
\arguments{
\item{data}{either a null object or a \code{\link[survival]{survfit}} object or a data frame that has
the variables \code{time} specifying time to event or time to censoring,
\code{status} specifying the censoring status and \code{group} specifying the
 different groups to be compared.}

\item{surv_KM}{either a null object or a numeric vector containing survival
rates for one or two groups}

\item{se_KM}{either a null object or a numeric vector containing standard
errors for one or two groups}

\item{t}{the point in time at which the survival rate will be calculated}
}
\value{
A list of class \code{\link[EnvStats:htest.object]{htest}} containing information about the
p-value, the value of the test statistic, survival rates etc.
}
\description{
This function tests two survival rates at a fixed point in time for equality.
This test can also be used in order to compare the survival rate of a data
 frame to some given numerical value. Therefore, there are three
possibilities for data input, (1) \code{data} with two groups and
null variables \code{surv_KM} and \code{se_KM}, (2) \code{data} with one group and two
numeric vectors \code{surv_KM} and \code{se_KM} of length 1, (3) a null variable \code{data} and
two numeric vectors \code{surv_KM} and \code{se_KM} of length 2. The test uses a clog-transformed
 test statistic for comparing the survival rates of two survival curves at a
 fixed point in time. This statistic is an implementation of the
test statistic \eqn{X_3^2}{X3^2} in Klein et al.'s paper [1]. The test is a
chi-squared test with one degree of freedom.
}
\examples{
data(exp_surv)
clog.test(exp_surv, t=1)

library(survival)
exp_fit <- survfit(formula = Surv(time, status) ~ group, data=exp_surv)
clog.test(exp_fit, t=1)

library(dplyr)
exp_surv \%>\% filter(group == 1) -> exp_surv_grp1
clog.test(exp_surv_grp1, surv_KM=0.5, se_KM=0.08, t=1)

clog.test(surv_KM=c(0.5, 0.6), se_KM=c(0.1,0.1))

}
\references{
[1]  Klein, J.P., Logan, B., Harhoff, M. and Andersen, P.K. (2007),
\href{https://doi.org/10.1002/sim.2864}{Analyzing survival curves at a fixed point in time}.
Statist. Med., 26: 4505-4519.
}
