% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_utils.R
\name{get_sigma_KM}
\alias{get_sigma_KM}
\title{Square root of Greenwood's formula}
\usage{
get_sigma_KM(surv_KM, se_KM)
}
\arguments{
\item{surv_KM}{a numeric vector containing survival rates for two groups}

\item{se_KM}{a numeric vector containing standard errors for two groups}
}
\value{
A numeric vector containing the square root of Greenwood's formula for
 the different groups.
}
\description{
This function returns the square root of the quotient of the Kaplan-Meier
estimate of the survival rate by its standard error. This quotient can be
calculated by a formula known as Greenwood's formula. This helper function is
simply introduced for better readability when comparing the implemented
functions to the ones from Klein et al.
}
\examples{
get_sigma_KM(c(0.8, 0.9), c(0.11, 0.09))

}
