% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_utils.R
\name{get_surv_KM}
\alias{get_surv_KM}
\title{Kaplan-Meier estimate of the survival rate}
\usage{
get_surv_KM(data = NULL, surv_KM = NULL, t = NULL)
}
\arguments{
\item{data}{either a null object or a \code{\link[survival]{survfit}} object or a data frame that has
the variables \code{time} specifying time to event or time to censoring,
\code{status} specifying the censoring status and \code{group} specifying the
 different groups to be compared.}

\item{surv_KM}{either a null object or a numeric vector containing survival
rates for one or two groups}

\item{t}{the point in time at which the survival rate will be calculated}
}
\value{
A numeric vector containing the survival rates for the different
groups at time t.
}
\description{
This function returns the Kaplan-Meier estimate of the survival rate at a
fixed point in time for two groups. There are three possibilities for data
input, (1) data with two groups and a null variables surv_KM,
(2) data with one group and a numeric vector surv_KM of length 1,
(3) a null variable data and a numeric vector surv_KM of length 2. In case
 (1), it is a wrapper function of the survival estimate from the \code{\link{survival}}
package. In case (2), it also uses this function, and then concatenates it
with the numeric vector surv_KM. In case (3), it simply returns the numeric
vector surv_KM
}
\examples{
data(exp_surv)
get_surv_KM(exp_surv, t=1)

library(dplyr)
exp_surv \%>\% filter(group == 1) -> exp_surv_grp1
get_surv_KM(exp_surv_grp1, surv_KM=0.5, t=1)

get_surv_KM(surv_KM=c(0.5,0.6))

}
