% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_utils.R
\name{get_surv_KM_from_data}
\alias{get_surv_KM_from_data}
\title{Kaplan-Meier estimate of the survival rate}
\usage{
get_surv_KM_from_data(data, t)
}
\arguments{
\item{data}{either a \code{\link[survival]{survfit}} object or a data frame that has the variables
\code{time} specifying time to event or time to censoring, \code{status}
specifying the censoring status and \code{group} specifying the different
groups to be compared.}

\item{t}{the point in time at which the survival rate will be calculated}
}
\value{
A numeric vector containing the survival rates for the different
groups at time t.
}
\description{
This function returns the Kaplan-Meier estimate of the survival rate at a
fixed point in time. In fact, it is a wrapper function of the survival
estimate from the \code{\link{survival}} package.
}
\examples{
data(exp_surv)
get_surv_KM_from_data(exp_surv, 1)

}
