% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_utils.R
\name{get_survfit}
\alias{get_survfit}
\title{Function for generating a survfit object.}
\usage{
get_survfit(data)
}
\arguments{
\item{data}{a data frame that has the variables \code{time} specifying time to
event or time to censoring, \code{status} specifying the censoring status and
\code{group} specifying the different groups to be compared.}
}
\value{
A \code{\link[survival]{survfit}} object.
}
\description{
This function returns a \code{\link[survival]{survfit}} object that has been generated from a
survival data set.
}
\examples{
data(exp_surv)
get_survfit(exp_surv)

}
